Modification for better use
In this commit the following changes has been made:
- All echos are removed
- Limit each line to only have 80 characters
- At the beginning, check if this script is run under the currect
  location, and if it is wrong, throw error and exit
- Change owner of /build and /madlib repo to postgres user, so that
  later postgres user can make changes and build MADlib

+++ tool/docker_start.sh
@@ -16,7 +16,7 @@
 # specific language governing permissions and limitations
 # under the License.
 
-#####################################################################################
+###############################################################################
 #  This is a script that does the following:
 #  * Pull madlib/postgres_9.6:latest from docker;
 #  * Mount your local madlib directory to docker container;
@@ -25,9 +25,9 @@
 #  * Log in to docker container from shell as postgres user, so that you
 #    can run psql and start using madlib from here
 #
-# Note: you have to run this script from one level up from your madlib directory, for
-# example, ~/workspace which has madlib src directory there
-#####################################################################################
+# Note: you have to run this script from one level up from your madlib 
+# directory, for example, ~/workspace which has madlib src directory there
+###############################################################################
 set -o pipefail
 
 workdir=$(pwd)
@@ -35,10 +35,18 @@ user_name=$(whoami)
 reponame=madlib
 
 echo "======================================================================"
-echo "Build user: $user_name"
-echo "Work directory: $workdir"
-echo "Git reponame: $reponame"
+echo "Build user: ${user_name}"
+echo "Work directory: ${workdir}"
+echo "Git reponame: ${reponame}"
 echo "======================================================================"
+
+if [[ ! -d "${workdir}/${reponame}" ]]; then
+  echo "Error: you have to run this script from one level up from your madlib \
+        directory, for example, ~/workspace which has madlib src directory \
+        there"
+  exit
+fi
+
 rm -rf logs
 mkdir logs
 
@@ -51,25 +59,35 @@ echo "Creating docker container"
 docker pull madlib/postgres_9.6:latest
 
 # Launch docker container with volume mounted from workdir
-docker run -d --name madlib -v "${workdir}/${reponame}":/madlib madlib/postgres_9.6 | tee logs/docker_setup.log
+docker run -d --name madlib -v "${workdir}/${reponame}":/madlib \
+ 					madlib/postgres_9.6 | tee logs/docker_setup.log
 
 ## This sleep is required since it takes a couple of seconds for the docker
-## container to come up, which is required by the docker exec command that follows.
+## container to come up, which is required by the docker exec command that 
+## follows.
 sleep 5
 
 echo "---------- Building MADlib -----------"
 # cmake, make, make install
-docker exec madlib bash -c 'rm -rf /build; mkdir /build; cd /build; cmake ../madlib; make; make install' | tee $workdir/logs/madlib_compile.log
+docker exec madlib bash -c "rm -rf /build; mkdir /build; cd /build; cmake \
+							../madlib; make; make install" \
+			| tee "${workdir}/logs/madlib_compile.log"
 
 echo "---------- Installing and running install-check --------------------"
 # Install MADlib and run install check
-docker exec madlib bash -c '/build/src/bin/madpack -p postgres -c postgres/postgres@localhost:5432/postgres install' | tee $workdir/logs/madlib_install.log
+docker exec madlib bash -c "/build/src/bin/madpack -p postgres -c \
+							postgres/postgres@localhost:5432/postgres install" \
+			| tee "${workdir}/logs/madlib_install.log"
 
 docker exec madlib bash -c 'mkdir -p /tmp'
-docker exec madlib bash -c '/build/src/bin/madpack -p postgres  -c postgres/postgres@localhost:5432/postgres -d /tmp install-check' | tee $workdir/logs/madlib_install_check.log
+docker exec madlib bash -c "/build/src/bin/madpack -p postgres  -c \
+							postgres/postgres@localhost:5432/postgres -d \
+							/tmp install-check" \
+			| tee "${workdir}/logs/madlib_install_check.log"
 
 # To run psql, you have to login as postgres, not root
 echo "---------- docker exec image as user postgres-----------"
 docker exec madlib bash -c 'chown -R postgres:postgres /build'
 docker exec madlib bash -c 'chown -R postgres:postgres /madlib'
-docker exec --user=postgres -it madlib bash | tee $workdir/logs/docker_exec.log
+docker exec --user=postgres -it madlib bash \
+		| tee "${workdir}/logs/docker_exec.log"
